build: Update to kstdlib-20190509
- Use the clang bundled in the kstdlib build for CC instead of the
  system version.

- Update to use the new location of swiftrt.o

- Add -ggdb flags when compiling C code.

- When building the ISO image, increase the size to 20M and stop
  adding images that are not needed. Use -e to specify the EFI boot
  image. This reduces the number of unnecessary files on the boot CD
  and allows for a larger kernel image.

- linker.script: Remove the .swift2_protocol_conformances and
  .swift2_type_metadata sections as these are unneeded now.

- Increase the stack size to 16 pages.

- Minor Swift 5 warning fixes.

+++ kernel/devices/apic.swift
@@ -281,7 +281,7 @@ public class APIC: InterruptController {
         let bootProcessor = apicStatus[bootProcessorBit] == 1
         let maxPhyAddrBits = CPU.capabilities.maxPhyAddrBits
         let lomask = ~(UInt(1 << 12) - 1)
-        let himask = (1 << maxPhyAddrBits) - 1
+        let himask = (UInt(1) << maxPhyAddrBits) - 1
         let mask = lomask & himask
         let address = RawAddress(apicStatus.toUInt64() & UInt64(mask))
         let baseAddress = PhysAddress(address)
